# synkdata/docs/api.yaml
openapi: 3.0.3
info:
  title: SynkData Identity Intelligence Platform API
  description: |
    Plataforma de inteligencia de identidad para verificación, antecedentes,
    cumplimiento normativo y análisis de riesgo enfocada en México y LATAM.

    ## Autenticación
    Todos los endpoints protegidos requieren un token JWT en el header
    `Authorization: Bearer <token>`.

    ## Códigos de respuesta
    - `200` Operación exitosa
    - `201` Recurso creado
    - `400` Error de validación en los datos de entrada
    - `401` Token de autenticación inválido o expirado
    - `403` Permisos insuficientes para la operación
    - `404` Recurso no encontrado
    - `429` Límite de peticiones excedido
    - `500` Error interno del servidor
  version: 1.0.0
  contact:
    name: SynkData Team
    email: dev@synkdata.io
    url: https://synkdata.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Servidor de desarrollo
  - url: https://api.synkdata.io/api/v1
    description: Servidor de producción
security:
  - BearerAuth: []
...
paths:
...
  /curp/{curp}:
    get:
      tags: [Verification]
      summary: Obtener validación cacheada de CURP
      operationId: getCachedCurpValidation
      security:
        - BearerAuth: []
      parameters:
        - name: curp
          in: path
          required: true
          schema:
            type: string
            minLength: 18
            maxLength: 18
          description: CURP a buscar en el caché
          example: "GOME850101HDFRRN09"
      responses:
        '200':
          description: Resultado de validación cacheado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurpValidationResult'
        '404':
          description: No se encontró una validación previa
...
  /rfc/{rfc}:
    get:
      tags: [Verification]
      summary: Obtener validación cacheada de RFC
      operationId: getCachedRfcValidation
      security:
        - BearerAuth: []
      parameters:
        - name: rfc
          in: path
          required: true
          schema:
            type: string
            minLength: 12
            maxLength: 13
          description: RFC a buscar en el caché
          example: "GOME850101ABC"
      responses:
        '200':
          description: Resultado de validación cacheado
        '404':
          description: No se encontró una validación previa

  # ── SCREENING ─────────────────────────────────────────────────────────
...
  /screening/entity:
    post:
      tags: [Screening]
      summary: Screening de persona moral o entidad corporativa
      description: |
        Realiza el screening de una persona moral o entidad corporativa
        contra las listas restrictivas aplicables.
      operationId: screenEntity
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScreeningEntityRequest'
            example:
              name: "Corporativo Alfa S.A. de C.V."
              country: "MX"
      responses:
        '200':
          description: Screening completado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreeningResult'
...
  /identity/correlate:
    post:
      tags: [Identity]
      summary: Correlación de identidad
      description: |
        Ejecuta la correlación cruzada de todas las señales de identidad
        proporcionadas (nombre, CURP, RFC, correo, teléfono, redes sociales)
        para determinar la consistencia y confianza de la identidad declarada.
      operationId: correlateIdentity
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityData'
            example:
              name: "JUAN GOMEZ RODRIGUEZ"
              curp: "GOME850101HDFRRN09"
              rfc: "GOME850101ABC"
              email: "juan.gomez@empresa.com.mx"
              phone: "+525512345678"
              username: "jgomez"
              company: "Empresa S.A."
              domain: "empresa.com.mx"
      responses:
        '200':
          description: Correlación completada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrelationResult'
  /identity/trust-score:
    post:
      tags: [Identity]
      summary: Cálculo del puntaje de confianza (Trust Score)
      description: |
        Calcula el puntaje de confianza de una identidad basándose en
        señales positivas: RENAPO válido (+20), RFC válido (+15),
        SAT activo (+15), screening limpio (+20), etc.
      operationId: calculateTrustScore
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrustContext'
      responses:
        '200':
          description: Cálculo completado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustScoreResult'
  /identity/assess:
    post:
      tags: [Identity]
      summary: Evaluación completa de identidad
      description: |
        Realiza la evaluación integral de identidad combinando:
        1) Correlación cruzada de señales, 2) Cálculo del trust score,
        3) Evaluación de riesgo con recomendación de decisión.
      operationId: assessIdentity
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityData'
            example:
              name: "JUAN GOMEZ RODRIGUEZ"
              curp: "GOME850101HDFRRN09"
              rfc: "GOME850101ABC"
              email: "juan.gomez@empresa.com.mx"
              phone: "+525512345678"
      responses:
        '200':
          description: Evaluación completada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullIdentityResponse'
  /identity/{assessment_id}:
    get:
      tags: [Identity]
      summary: Consultar evaluación de identidad previa
      operationId: getIdentityAssessment
      security:
        - BearerAuth: []
      parameters:
        - name: assessment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Identificador único de la evaluación
      responses:
        '200':
          description: Evaluación encontrada
        '404':
          description: Evaluación no encontrada

  # ── RISK ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido del endpoint de autenticación.
        Formato: `Authorization: Bearer <token>`
  schemas:
...
    ScreeningEntityRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 500
          description: Nombre de la entidad corporativa
          example: "Corporativo Alfa S.A. de C.V."
        country:
          type: string
          maxLength: 5
          default: "MX"
          description: País de registro
        threshold:
          type: number
          minimum: 0.5
          maximum: 1.0
          description: Umbral de similitud personalizado
...
    IdentityData:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Nombre completo de la persona
          example: "JUAN GOMEZ RODRIGUEZ"
        curp:
          type: string
          description: CURP de la persona
          example: "GOME850101HDFRRN09"
        rfc:
          type: string
          description: RFC de la persona
          example: "GOME850101ABC"
        email:
          type: string
          format: email
          description: Correo electrónico
          example: "juan.gomez@empresa.com.mx"
        phone:
          type: string
          description: Número telefónico (formato E.164 preferido)
          example: "+525512345678"
        username:
          type: string
          description: Nombre de usuario
          example: "jgomez"
        company:
          type: string
          description: Empresa donde labora
          example: "Empresa S.A."
        domain:
          type: string
          description: Dominio de la empresa
          example: "empresa.com.mx"
        social_profiles:
          type: array
          description: Perfiles en redes sociales
          items:
            type: object
            properties:
              platform:
                type: string
                description: Plataforma social
                example: "linkedin"
              url:
                type: string
                format: uri
                description: URL del perfil
                example: "https://linkedin.com/in/jgomez"
    TrustContext:
      type: object
      properties:
        renapo_valid:
          type: boolean
          default: false
          description: CURP validada contra RENAPO
        rfc_valid:
          type: boolean
          default: false
          description: RFC con formato y dígito verificador válidos
        sat_active:
          type: boolean
          default: false
          description: RFC activo en el SAT
        screening_clean:
          type: boolean
          default: false
          description: Screening en listas restrictivas sin coincidencias
        professional_presence:
          type: boolean
          default: false
          description: Presencia profesional verificable
        github_active:
          type: boolean
          default: false
          description: Perfil activo en GitHub
        linkedin_found:
          type: boolean
          default: false
          description: Perfil encontrado en LinkedIn
        email_verifiable:
          type: boolean
          default: false
          description: Correo electrónico verificable (MX + SMTP)
        phone_valid:
          type: boolean
          default: false
          description: Número telefónico válido
...
    CurpValidationResult:
      type: object
      properties:
        is_valid:
          type: boolean
          description: Si la CURP pasó todas las validaciones
        format_valid:
          type: boolean
          description: Si el formato de la CURP es correcto
        check_digit_valid:
          type: boolean
          description: Si el dígito verificador es correcto
        renapo_match:
          type: boolean
          nullable: true
          description: Si la CURP fue encontrada en RENAPO
        extracted_info:
          type: object
          nullable: true
          description: Información extraída de la CURP
          properties:
            name_initials:
              type: string
              example: "GOME"
            birth_date:
              type: string
              example: "1985-01-01"
            gender:
              type: string
              example: "H"
            state_code:
              type: string
              example: "DF"
            state_name:
              type: string
              example: "Ciudad de México"
            internal_consonants:
              type: string
              example: "RRN"
        errors:
          type: array
          items:
            type: string
          description: Lista de errores encontrados
        validated_at:
          type: string
          format: date-time
...
    ScreeningResult:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        matches:
          type: array
          items:
            $ref: '#/components/schemas/MatchDetail'
        total_hits:
          type: integer
          minimum: 0
        max_score:
          type: number
          minimum: 0.0
          maximum: 1.0
        risk_level:
          type: string
          enum: [none, low, medium, high, critical]
        sources_checked:
          type: array
          items:
            type: string
        sources_failed:
          type: array
          items:
            type: string
        timestamp:
          type: string
    MatchDetail:
      type: object
      properties:
        source:
          type: string
          description: Fuente donde se encontró la coincidencia
          example: "ofac"
        score:
          type: number
          minimum: 0.0
          maximum: 1.0
          description: Puntuación de similitud
          example: 0.95
        match_type:
          type: string
          enum: [exact, fuzzy, phonetic, alias]
          description: Tipo de coincidencia
        entity_name:
          type: string
          description: Nombre de la entidad coincidente
          example: "JUAN GOMEZ RODRIGUEZ"
        entity_data:
          type: object
          additionalProperties: true
          description: Datos adicionales de la entidad
        is_confirmed:
          type: boolean
          description: Si la coincidencia fue confirmada manualmente
...
    CorrelationResult:
      type: object
      properties:
        identity_confidence:
          type: number
          minimum: 0.0
          maximum: 100.0
          description: Puntaje de confianza de identidad (0-100)
          example: 85.5
        signals:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              passed:
                type: boolean
              details:
                type: string
        warnings:
          type: array
          items:
            type: string
        flags:
          type: array
          items:
            type: string
    TrustScoreResult:
      type: object
      properties:
        score:
          type: number
          minimum: 0.0
          maximum: 100.0
          description: Puntuación de confianza (0-100)
          example: 85.0
        max_possible:
          type: number
          example: 100.0
        level:
          type: string
          enum: [very_high, high, medium, low, very_low]
          description: Nivel de confianza categorizado
        contributors:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              points:
                type: number
              max_points:
                type: number
              passed:
                type: boolean
              details:
                type: string
    RiskAssessmentResult:
      type: object
      properties:
        risk_score:
          type: number
          minimum: 0.0
          maximum: 100.0
          description: Puntuación de riesgo (0-100)
          example: 15.0
        trust_score:
          type: number
          minimum: 0.0
          maximum: 100.0
          description: Puntuación de confianza
          example: 85.0
        recommendation:
          type: string
          enum: [APPROVE, REVIEW, REJECT]
          description: Recomendación de decisión
        risk_factors:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              score:
                type: number
              severity:
                type: string
                enum: [critical, high, medium, low]
              details:
                type: string
        mitigating_factors:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              points_reduced:
                type: number
              details:
                type: string
    FullIdentityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        identity_data:
          $ref: '#/components/schemas/IdentityData'
        correlation:
          $ref: '#/components/schemas/CorrelationResult'
        trust_score:
          $ref: '#/components/schemas/TrustScoreResult'
        risk_assessment:
          $ref: '#/components/schemas/RiskAssessmentResult'
        evaluated_at:
          type: string
          format: date-time
...